+++ src/com/Rit.java
@@ -1,7 +1,7 @@
 package com;
 
 public class Rit extends Stap{
-    public Rit(Node naar, int kilometers) {
+    public Rit(Node naar, double kilometers) {
         super(naar, kilometers);
     }
 }
